code/requirements.txt
Django==4.2.2
gunicorn==20.1.0
gevent==22.10.2
psycopg2==2.9.6
django-registration==3.3
requests==2.31.0
django-admin-tools==0.9.2
